Changed NOMINAL.new to have the old behaviour. Added NOMINAL.newBasedOn to generate a name based on name.
There are mainly two scenarious when we want to generate a fresh name: (a) when alpha renaming, (b) when generating a placeholder.
For the scenario (a), we want to be able to compare with the old term, thus the generated names should resemble the old wones (newBasedOn). For (b), any name would do (new).

This change allowed to get rid of the ugly requirement on the symbolic instance -- nameOfConstrSort.

git-svn-id: svn+ssh://fries.it.uu.se/home/victor/SVNroot/pa/gepi/trunk/PsiWorkBench/trunk@3978 15178899-d6b2-c17f-aa0b-88a7dc59b424